Improvement in docker_backup_tests.py
