+++ test_scripts/pxb/docker_backup_tests.py
@@ -239,35 +239,6 @@ def wait_for_mysql_container(container_name, max_wait=180):
     return True
 
 
-# Pytest configuration
-def pytest_addoption(parser):
-    """Add command-line options for pytest."""
-    parser.addoption(
-        "--repo-name",
-        action="store",
-        default=None,
-        help="Repo name: pxb-24, pxb-80, pxb-8x-innovation, pxb-84-lts, pxb-9x-innovation"
-    )
-    parser.addoption(
-        "--repo-type",
-        action="store",
-        default=None,
-        help="Repo type: release, testing, experimental"
-    )
-    parser.addoption(
-        "--server",
-        action="store",
-        default=None,
-        help="Server: ps, ms"
-    )
-    parser.addoption(
-        "--innovation",
-        action="store",
-        default="",
-        help="Innovation version: 8.1, 8.2, 8.3, 8.4, 9.1"
-    )
-
-
 @pytest.fixture(scope="function")
 def test_config(request):
     """Fixture to get test configuration from command-line options."""
